+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/CompositeMediaSource.java
@@ -28,6 +28,7 @@ import androidx.media3.common.util.Util;
 import androidx.media3.datasource.TransferListener;
 import androidx.media3.exoplayer.drm.DrmSession;
 import androidx.media3.exoplayer.drm.DrmSessionEventListener;
+import androidx.media3.exoplayer.drm.KeyRequestInfo;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Objects;
@@ -320,9 +321,10 @@ public abstract class CompositeMediaSource<T> extends BaseMediaSource {
     }
 
     @Override
-    public void onDrmKeysLoaded(int windowIndex, @Nullable MediaPeriodId mediaPeriodId) {
+    public void onDrmKeysLoaded(
+        int windowIndex, @Nullable MediaPeriodId mediaPeriodId, KeyRequestInfo keyRequestInfo) {
       if (maybeUpdateEventDispatcher(windowIndex, mediaPeriodId)) {
-        drmEventDispatcher.drmKeysLoaded();
+        drmEventDispatcher.drmKeysLoaded(keyRequestInfo);
       }
     }